+++ src/ann/__init__.py
@@ -1,3 +1,3 @@
 """Ann core package."""
 
-__version__ = "0.0.1"
+__version__ = "0.0.2"